Refactor ArgumentParser
