+++ test/domain/arguments/argument_parser_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:nyrna/domain/arguments/argument_parser.dart';
+import 'package:nyrna/main.dart';
 
 void main() {
   group('ArgumentParser:', () {
